+++ plugins/events-backend/src/index.ts
@@ -20,6 +20,5 @@
  * @packageDocumentation
  */
 
-export { EventsBackend } from './service/EventsBackend';
+export * from './deprecated';
 export { HttpPostIngressEventPublisher } from './service/http';
-export { DefaultEventBroker } from './service/DefaultEventBroker';