+++ plugins/kubernetes-backend/src/service/KubernetesBuilder.ts
@@ -64,7 +64,12 @@ import {
 } from './KubernetesFanOutHandler';
 import { KubernetesClientBasedFetcher } from './KubernetesFetcher';
 import { KubernetesProxy } from './KubernetesProxy';
-import { TokenManager } from '@backstage/backend-common';
+import { createLegacyAuthAdapters, TokenManager } from '@backstage/backend-common';
+import {
+  AuthService,
+  DiscoveryService,
+  HttpAuthService,
+} from '@backstage/backend-plugin-api';
 
 /**
  *
@@ -74,8 +79,11 @@ export interface KubernetesEnvironment {
   logger: Logger;
   config: Config;
   catalogApi: CatalogApi;
+  discovery: DiscoveryService;
   permissions: PermissionEvaluator;
   tokenManager: TokenManager;
+  auth?: AuthService;
+  httpAuth?: HttpAuthService;
 }
 
 /**
@@ -133,6 +141,13 @@ export class KubernetesBuilder {
         router: Router(),
       } as unknown as KubernetesBuilderReturn;
     }
+
+    const { auth, httpAuth } = createLegacyAuthAdapters({
+      auth: this.env.auth,
+      httpAuth: this.env.httpAuth,
+      discovery: this.env.discovery,
+    });
+
     const customResources = this.buildCustomResources();
 
     const fetcher = this.getFetcher();
@@ -160,6 +175,8 @@ export class KubernetesBuilder {
       this.env.catalogApi,
       proxy,
       permissions,
+      auth,
+      httpAuth,
     );
 
     return {
@@ -340,6 +357,8 @@ export class KubernetesBuilder {
     catalogApi: CatalogApi,
     proxy: KubernetesProxy,
     permissionApi: PermissionEvaluator,
+    authService: AuthService,
+    httpAuth: HttpAuthService,
   ): express.Router {
     const logger = this.env.logger;
     const router = Router();
@@ -394,7 +413,13 @@ export class KubernetesBuilder {
       });
     });
 
-    addResourceRoutesToRouter(router, catalogApi, objectsProvider);
+    addResourceRoutesToRouter(
+      router,
+      catalogApi,
+      objectsProvider,
+      authService,
+      httpAuth,
+    );
 
     return router;
   }