+++ plugins/scaffolder-backend-module-cookiecutter/src/actions/fetch/cookiecutter.test.ts
@@ -14,19 +14,15 @@
  * limitations under the License.
  */
 
-import {
-  getVoidLogger,
-  UrlReader,
-  ContainerRunner,
-} from '@backstage/backend-common';
+import { UrlReader, ContainerRunner } from '@backstage/backend-common';
 import { ConfigReader } from '@backstage/config';
 import { JsonObject } from '@backstage/types';
 import { ScmIntegrations } from '@backstage/integration';
 import { createMockDirectory } from '@backstage/backend-test-utils';
-import { PassThrough } from 'stream';
 import { createFetchCookiecutterAction } from './cookiecutter';
 import { join } from 'path';
 import type { ActionContext } from '@backstage/plugin-scaffolder-node';
+import { createMockActionContext } from '@backstage/plugin-scaffolder-node-test-utils';
 
 const executeShellCommand = jest.fn();
 const commandExists = jest.fn();
@@ -88,7 +84,7 @@ describe('fetch:cookiecutter', () => {
   beforeEach(() => {
     jest.resetAllMocks();
 
-    mockContext = {
+    mockContext = createMockActionContext({
       input: {
         url: 'https://google.com/cookie/cutter',
         targetPath: 'something',
@@ -101,11 +97,7 @@ describe('fetch:cookiecutter', () => {
         baseUrl: 'somebase',
       },
       workspacePath: mockTmpDir,
-      logger: getVoidLogger(),
-      logStream: new PassThrough(),
-      output: jest.fn(),
-      createTemporaryDirectory: jest.fn().mockResolvedValue(mockTmpDir),
-    };
+    });
     mockDir.setContent({ template: {} });
 
     commandExists.mockResolvedValue(null);