+++ plugins/scaffolder-node-test-utils/src/actions/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2020 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,4 @@
  * limitations under the License.
  */
 
-// Update with your config settings.
-
-// This file makes it possible to run "yarn knex migrate:make some_file_name"
-// to assist in making new migrations
-module.exports = {
-  client: 'better-sqlite3',
-  connection: ':memory:',
-  useNullAsDefault: true,
-  migrations: {
-    directory: './migrations',
-  },
-};
+export { createMockActionContext } from './mockActionConext';