+++ plugins/apache-airflow/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-apache-airflow
 
+## 0.2.21-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.14.1-next.0
+  - @backstage/core-plugin-api@1.9.1-next.0
+
 ## 0.2.20
 
 ### Patch Changes