+++ plugins/github-deployments/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-github-deployments
 
+## 0.1.62-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/errors@1.2.4-next.0
+  - @backstage/core-components@0.14.1-next.0
+  - @backstage/integration-react@1.1.25-next.0
+  - @backstage/plugin-catalog-react@1.10.1-next.0
+  - @backstage/catalog-model@1.4.5-next.0
+  - @backstage/core-plugin-api@1.9.1-next.0
+  - @backstage/integration@1.9.1-next.0
+
 ## 0.1.61
 
 ### Patch Changes