+++ plugins/org-react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-org-react
 
+## 0.1.20-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.14.1-next.0
+  - @backstage/plugin-catalog-react@1.10.1-next.0
+  - @backstage/catalog-client@1.6.1-next.0
+  - @backstage/catalog-model@1.4.5-next.0
+  - @backstage/core-plugin-api@1.9.1-next.0
+
 ## 0.1.19
 
 ### Patch Changes