+++ plugins/tech-radar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-tech-radar
 
+## 0.6.14-next.0
+
+### Patch Changes
+
+- a2327ac: Fixed an issue with the "moved in direction" table header cell getting squished and becoming unreadable if a timeline description is too long
+- Updated dependencies
+  - @backstage/core-components@0.14.1-next.0
+  - @backstage/core-plugin-api@1.9.1-next.0
+  - @backstage/frontend-plugin-api@0.6.1-next.0
+  - @backstage/core-compat-api@0.2.1-next.0
+
 ## 0.6.13
 
 ### Patch Changes